Streaming Entity management fix, Mark I. - extended GTA:SA streaming entity garbage collection service, due to a limitation to 1000 streaming entities, the engine could not display more than 1000 objects, buildings and dummies at once. by removing that limit, GTA:SA displays very clearly - reverse engineered some world functions (CWorldSA) - fixed my archive loader - localized the CRunningScript hook into game_sa - improved list macros (rwlist.hpp) - improved CCameraSAInterface (offsets realigned, pls verify) - fixed animation loading bugs (some incompatibilities resolved)
What is left to do is break the limit of actual rendering. As of now, the limit is 65536 for different object types. That can be problematic in the future, because we can exhaust this (theoretically) exhaust it.
I need to reimplement procedual object collision. Should be there next commit.
We should test MTA:Eir together again.